requirements.txt
Pillow==2.9.0
pymongo==3.0.3
PySocks==1.5.6
pytz==2015.6
PyYAML==3.11
requests==2.8.1
romkan==0.2.1
selenium==2.47.3
wheel==0.24.0
